Add: aes keychain reading, writing, gen, business
big commit, im sorry i forgot :(. aes keychain reading and writing logic
is all here. generation and business logic is all here too.

our keychain is defined as a map of the keys to their extensions that
is signed and then encrypted by an RSA private key.

because of that, I chose json as the format, presigned and encrypted,
itll look like
```json
{
    "<extension>": "<base64 aes key>",
}
```

**writing**
the logic of writing is
* marshal key map to json
* sign that json(using md5, trust i didnt want it either)
* append json to end of signature
* encrypt whole payload with public key
* write to file
this will error out if file already exists, you can only have one keychain

**reading**
logic is basically the same as above except its in reverse
* read from file
* decrypt payload using private key
* verifies signature using public key
* unmarshals json to map

The business logic for aes keys is try to load them in, if dont exist
gen.