Add: encrypt concurrent walking, Fix: gitignore
gitignore had `encryptdir` in it, whoops. encrypt concurrent walking is
a wild time. It is fully concurrent and goes **zoom**.

There is a hidden race condition and Ill talk about how I solved it.

The race condition is if the same directory is gotten to by the walker
in different workers, it will encrypt a file multiple times. Thats no
bueno because its concurrent, no guarantee about run order or
syncronization.

**How we fix this**
we sign the hash of the AES key and put that at the beginning of the
encrypted file. if it doesnt exist when reading the file, were safe to
act on it, else weve already encrypted this file.

This doesnt solve the case where two workers get to the same file around
the same time. To solve that, we create a new file `<path>.enc` and
transfer and encrypt from `path` to `path.enc`, this also gives us the
ability to do an exclusive `open` call. That is atomic so we don't have
to worry about the case that its successful both time, the OS guarentees
it for us(enough unless people are writing shitty drivers).