Add: decrypt concurrent walker
basically the inverse of the encrypt one.

we create `path` + ".dec", if it exists then skip because another
goroutine is using it. then verify signature, if it isnt verified then
skip cause its either been decrypted or is corrupted. then read the
original message size. decrypt, trim padding size, and thats it.

the other stuff is literally copy pasta from the encryption one.