Fix: padding issues and also some other bugs
so there was a a few bad nasty bugs that caused padding issues. one was
reading a buffer for signature that was signature sized. if the file was
smaller, it had a bunch of nulls and caused issues. I also was saving
the wrong order for origSized writing after the padding happens,
obviously caused issues.